requirements.txt
faster-whisper>=1.0.0
openai>=1.0.0
sounddevice>=0.4.6
pynput>=1.7.6
pyperclip>=1.8.2
python-dotenv>=1.0.0
numpy>=1.24.0
